+++ lib/src/_core/utils/omni_video_player_initializer_factory.dart
@@ -32,7 +32,7 @@ class OmniVideoPlayerInitializerFactory {
 
     switch (sourceType) {
       case VideoSourceType.youtube:
-        if (sourceConfig.forceYoutubeWebViewOnly) {
+        if (sourceConfig.youtubeWebView.forceWebViewOnly) {
           return YouTubeWebViewInitializer(
             config: config,
             globalController: globalController,