+++ lib/src/utils/overlay_transition_switcher.dart
@@ -52,7 +52,7 @@ class OverlayTransitionSwitcher extends StatelessWidget {
         fit: fit,
         children: <Widget>[
           ...previousChildren,
-          if (currentChild != null) currentChild,
+          ?currentChild,
         ],
       ),
     );